Set baseUri to point to frontend
Setting it to point to backend was a bit premature, since currently
it's the frontend that's the only public facing part. In the future,
the idea is that the API will through content negotiation serve both
JSON-LD and HTML, making it in essence both an API and a web page.