Filter out dangling bnodes in external refs
We treat any reference to something that doesn't exist in the document
as an external reference. Bnodes are only internal references, but in
some documents there are dangling references (i.e. a reference to a
bnode that is no longer part of the document). This change filters out
these from the external references.