Reimplement OAIPMH bib:[location] set queries without CTEs or subquerys
The previous implementation of bib:[location] sets was correct, but the SQL it used was not efficient enough to work on scale.
This commit changes that by using a combination of JOINs instead of CTEs and subquerys.
The downside is that it may no longer be possible to list bib:[location] sets in the header for each document.
That feature has at least for now been disabled, meaning we're no longer 100% OAIPMH spec compliant.