Add a RefreshAPI, which performs needed side effects after changing a record.
When a record is changed through some non-standard measuse (for example by
hand), all of the derivative data for that record (Elastic mirror image, and
id/dependency tables) must also be updated. This API performs these update on
any IDs passed to it. This can be done either loudly (updating the records
modified timestamp) or quietly as chosen by the operator.