Add mechanism to prevent concurrent writing of duplicate holding posts
If two requests to add holding posts for a given library and bibliographic
post arrives, only one must result in an actual write. Multiple holdings
for the same bib/sigel are not permitted.

This commit adds an explicit row lock that can be taken in PostgreSQL component.
It then adds the taking of this lock on the linked bib post when writing holding
posts. While under this lock the existence of any conflicting holding posts is
determined.