Make searches more intelligent with @types.
The list of types that are used to filter a search will now first be
pruned of any type that has a subtype in the same list. This way a
search on for example [Instance, Electronic] is narrowed to a search
on only [Electronic].

After the narrowing is completed, all subclasses of the remaining types
are added to the search aswell, like before.