Make database deadlocks and connection leaks impossible.
This commit adds tracking of connection allocations for every thread
and logs an error if a connection is leaked.

It also splits the connection pool into two segments.

When a thread gets an unnested connection, only the first segment from
the pool is available to choose from.
However, if a call to getConnection() is made before the previous
connection has been released (a nested connection), the second segment
will also be available. This ensures that deadlocks where "everyone"
is holding their fist connection and waiting for their second (which
never becomes available) can't happen.

Attempting to get a third nested connection is not allowed (and results
in an immediate error). If this is deemed necessary in the future it
can be achieved by segmenting the pool further.