Make the batch import use a more cautious strategy for merging/enriching.
Ideally, this should be possible to loosen later, but for the initial
launch, the enriching should use RULE_PREFER_ORIGINAL as the default
rule (as opposed to RULE_AGGREGATE).

Additionaly, use RULE_AGGREGATE explicitly for all termins in the
forced_set_termslist.

The point of all this is enforce MARC repeatability when enriching
records. Hopefully we can allow for more permissive enriching later.