Improve revert prevention of multiple fields from one entity
Still rough, but fixes the worst bug of dropping record fields since the
top level entity record was tagged as reverted by the first field with
data about it!

Added a spec for this using a combo of 035 and 040.

(This also teased out some mapping problems in auth, e.g. when the thing
has a marc:fullerFormName or a marc:partNumber, a revert produces more
fields from those.)