Make initialize more lazy and add term validation
To allow for the JsonLd utility to be set later than construction
and still be usable when initializing from the config, run
initialize when JsonLd is set, and also conditionally before needed
(in case of running without the utility).

Then use it to validate the presence of a term in the vocabulary
upon initializing.