Work around overloaded use of repeatable in MarcSimpleFieldHandler
When addLink was present, use of repeatable was interpreted in
repeatable check also for property. (That specifically caused a spurious
warning for a column in bib 007 for Microform addLink: reductionRatio
with property: value.)

Also added an assert to ensure that addLink and addProperty aren't used
together (until this is potentially improved).