Block changing or removing an existing itemOf relation.
If we allow changing an itemOf relation to reference bib-B
instead of bib-A, the result is that bib-A will not be subject
to export when this action occurs. This in turn means that
websök and local systems will never be told that a holding
record for bib-A has disappeared.

This commit blocks this behaviour, but this implicitly also
blocks all merging of two existing records ("bibliographic overlay")
which tries to move holding records from one bib to another.

Mid-term this can be solved by changing bib overlay to delete
and recreate holdings instead of moving them.
Long-term this must be solved by altering the export-selection
mechanism to something more flexible than modified timestamps.