Improve move-lddb.sh script
Fix sed expression to rewrite both JSON-LD identifiers and column value
(in lddb__identifiers; won't work on first column since this matches on
tab character).

Also added maneuvers for invoking both pg_dump and psql from source to
destination. Change this if we want to do this in multiple steps.