Fix the research name form problem.
The problem could already be sidestepped by having authtype
set to to anything other than 'none', but the need for that
was only due to unnecessary (and damaging) attempts to optimize
the number of requests needed for each export in the export program,
where this code originated. The optimization was removed in the
export program earlier, but remained here, causing trouble with the
name forms.