Re-enable groupId for inseries-related data
This is necessary in order for both bib 490 and 8xx to be produced on
revert from a single SeriesRelationship-node containing both a
seriesStatement and an inSeries relation.

We held off doing this kind of merge of 440+490+8xx-data during the
initial conversion, since records were discovered for which more
intelligent merging appeared necessary.

This re-enabled "dumb" merge logic is deemed OK now in production since
the data continuously imported doesn't have those irregularities.