Export all affected bib records when a holding record is relinked multiple times.
When a holding record is relinked multiple times within an export-interval
we would previously only export the newly linked bib and the previously linked
bib. That is not enough, since a holding may have been relinked (to different
bibs) more than once during an interval.

The new approach is to export all bibs to which the holding _has_ been linked
within the update interval (including the bib it was linked to before the inteval).

As an implementation detail the initial scan for changes is now done on the
versions table instead of lddb, which means that exporting from time A to B
will always produce the same result, regardless of whether or not the records
that changed within that interval are changed again after the interval.