Merge publication objects on fuzzy dates
This compares date (as well as startYear plus optional endYear) with
year for fuzzy equivalence when merging a PrimaryPublication. We do so
by removing defined "fuzz" from the fuzzy value, and then remove the
stricter dates as well. If the result is empty, the fuzzy date is
considered equivalent enough.