Add script to add encodingLevel: marc:FullLevel, where it is missing
This global update of records is needed due to a incorrect
mapping, at the time of LibrisXL was released. All records that were
cataloged as marc:FullLevel, did not get any encodingLevel at all.

To correct this, the script adds encodingLvel: marc:FullLevel to all
records that:
* were created before LibrisXL was released
* are imported FullLevel records (specified in technicalNote)

NB. This script only takes care of FullLevel records.
Another script is needed to handle the remaining records with
missing encodingLevel.