Move script to scripts/2019/01 and remove one of the conditions
Remove condition for biblography.sigel, since it is not enough
reliable to say that all of these are fullLevel records. Analysis
of data has shown that there are examples of records that contains
this entity, but which still is not a fullLevel record. Futher
analysis are needed to determine which encodingLevel these records
should have.