Add support for "supplementary" subfields to marcframe
This is used as an alternative to "required" subfields where there is
more than one subfield that can be used for a "complete" field. If so,
the "supplementary" subfields are those which aren't enough in
themselves in order for a field to be considered complete to produce.

Initially this is used for bib 336, 337 and 338, since these could
otherwise be spuriously produced from a typeNote now, due to the
previous change.