Implement silentRevert and use for bib 041
This lets a field revert entities without marking them as reverted. That
allows for other fields to revert from the same entity.

(This is akin to the groupId mechanism, but allows for other fields to
be more exclusive. (E.g. bib 700[t] and 730 are disjoint (the former is
preferred over the latter for the same entity), but a repeated 041 picks
language from either of them without marking as reverted).)