Add handling of silentRevert for fixed field columns
Fixed columns by default do a silentRevert (i.e. do not mark entities
with _revertedBy). Now this can be controlled with a flag.

This flag is set to false for genreForm and carrierType. That avoids
values from them being simultaneously reverted to regular bib 655.

With this feature, some brittle handling cropped up regarding when a
field is considered to have failedRequired. This addresses that as well,
which in turn required adding required to some subfields in marcframe (a
seemingly sensible addition).

Two additional checks using the new succeededCodes introduced here have
also been added, to avoid a potential bug which can drop a field when a
second group of subfields (pertaining to the aboutNew mechanism) do not
produce a required field, even if it has already been produced in the
first group for the same field (hence passing the required rule).