Add boost computed from display lenses
Take the given type parameter into account when computing the boost
properties.

Currently controllable using a '_boost' parameter (for testing), with
options 'hardcoded' or a custom boost spec.

(Note that this is a complex approach which may be removed in favour of
a direct configuration, given sufficient testing and measurements of
desired results.)