Fix broken record mainEntity references when renaming the mainEntity
When LinkFinder.replaceSameAsLinksWithPrimaries is used, the mainEntity
reference is replaced by any already existing primary id. This collides
with purposefully changing the primary id, which is therefore restored
again by restoreNewCanonicalMainEntityUri.

(This is only relevant for the top entities of the graph, not any
eventually nested bnodes, which is why this restore is done instead of
conditionals within replaceSameAsLinksWithPrimaries.)