Clear out old code.
Much of this code was historically critical, but is today dead code at best.
In many cases, the code being removed no longer even functions when tried.

This commit essentially gets rid of all code intended to load XL from vcopy.

It also removes a couple of services that are no longer used or useful.
A "VcopyImporter servlet", and an APIX-client for example.