Regenerate the dependencies table after all documents have been copied.
When copying data from another whelk (filling a local or dev environment),
Each record is saved on its own, in sequence. When it is saved it writes its
dependencies in the dependencies table, but it cannot write dependencies to
records which have not already been written.

The result of this was that on the small environments the lddb__dependencies
table had incomplete data.

This change regenerates the lddb__dependencies table once all records are in
place to fix this.