Make records in Libris have the equivalent of 003 set to SE-LIBR.
Previously, when we imported a MARC record from elsewhere, we left
the record's 003 (@graph,0,identifiedBy,N,{@type=SystemNumber}), and
only changed 003 when exporting the record. The practical consequence
of this is that whatever ID was in 001 when we imported the record, it
was discarded instead of placed in 035$a to facilitate future matching
in other databases.

This is now being corrected, so that pre-import IDs end up in 035$a
and 003 is being set immediately on import.