Do not fail to remove recently changed records from Elastic.
When a record has (very) recently been updated, it cannot be deleted
until Elastic has decided that it (internally) is done with it, even though
it already answered that it was earlier. When this happens you get a
conflicting version-error. When such an error is registered, we should
retry - until the delete is successful.