Add known irregular terms to validator statistics report
Specifically, collect all terms that are sometimes, but not always used
as repeatable (set-container) terms from the MarcFrameConverter, and
list all paths to these.

Also report full paths on validation errors, and mark those that are
known as problematic (from MarcFrameConverter as per above).

The paths are also annotated with "1", "2+" or "5+" to represent one,
more than one or more than four list items respectively.