Remove the card cache.
The card cache serves an extremely marginal purpose (at best),
now that there is another cache for the completed results
of an embellishment (lddb__embellished).

It also has been known to have problematic race properties, and
(most importantly) it further complicates the cache invalidation
necessary when external data is to be thrown into the mix (not
just our own).

For these reasons, the card cache is being removed.