Fix logic bug, for OAIPMH silent-change publishing
The bug was that closed interval checks that included silent changes
would incorrectly accept records where one of the times was before the
end of the interal and _the other_ after the start of the interval.

For this to work, _one_ of the two supplied points in time must fall
within the interval. Not the combination of times.