Initial attempt to switch from getConnection to withDbConnection()
The purpose of this, is to make it impossible to end up in connection
deadlocks. The way it works is that you declare a withDbConnection
scope like so:

withDbConnection {
..
}

Within that scope you can call getMyConnection() as often as you like,
and you'll always be given the _same_ connection (never a new one).
Even when several withDbConnection blocks are nested, they will still
always give you that one same connection.
When the (outermost) scope ends, the connection is automatically released.
The idea is that this way, you can call getMyConnection() without fear
of the thread you're on already having reserved a connection (because that
would risk a deadlock).