Add a Java AutoCloseable version of the withDbConnection machinery.
You can, sort of, use the Groovy/Closure version from Java as well,
but in practice you'll be forced to use a lambda, which brings the
usual unfortunate consequences: Anything that throws a checked
Exception becomes extremely aggravating, and anything touched within
the lambda must first be declared final and so on.

Both of these turned out to make this impractical for our needs.

What happens now, is that you can instead use an AutoCloseable
(try-with-resources) version, in Java, which uses the same piece of
ThreadLocal memory as the Closure version. This means that the two
can be mixed freely.

In other words it is perfectly fine to do:
try (..ConnectionContext..) { <-- Java version
  getMyConnection() <-- Get same connection
  withDbConnection { <-- Groovy version
    getMyConnection() <-- Get same connection
  }
} <-- Implicit release