Don't fail to generate reverse links if thing IRI missing for one doc (#998)
At the moment of writing `lddb__dependencies` contains 62 ids
where the document is deleted. This has to be investigated
and fixed. Since they are deleted all IRIs are gone from
`lddb__identifiers`. This fix makes it so that we don't
drop all other `@reverse` links when one of these is among
the incoming links.