Make embellished responses cachable
Make ETag/If-None-Match (304 NOT MODIFIED) work with embellish.

(ETag is an opaque string so we can put whatever we want inside)
When `embellish==false` ETag is checksum of document
When `embellish==true` include checksum of both un-embellished and embellished document in ETag.

Both ETag forms can be used in PUT.