This fixes a bug where a scripted record change was seen as manual.
This bug/fix also applies to the batch import merge-logic, which is
based on the same history/ownership calculation.

The bug essentially is that scripted changes _used to_ set the
changedby column in a very specific way, which is no longer true.
The fix is to instead classify an edit as scripted/not-scripted
based on comparing modified and @graph,0,generationDate.