Use suffix or conneg when determining redirects
This adds the data portion to the target URL of the redirection to aid
clients in final retrieval. This is necessary for suffix based content
negotiation, since the client might not have sent any "Accept" header.

It is not necessary for "Accept"-header based content negotiation, but
apart from the *hinted* canonical URL, doing this appears to be OK.