Use the correct sorts of paths in the merge rule/ownership lookups.
"Magic paths" are those that (potentially) contain for example @type=PrimaryContribution
indexes instead of numeric list indexes (like "0").

"True paths" instead contain the normal numeric list indexes.

When merging, the rules must be looked up using the _magic_ paths, because that is the
form used in the rule DSL. However: the entity ownerships must _not_ be looked up that
way, they must instead be looked up using true paths, because that is (and must be) the
form used by the history generation.

Before this commit, the ownerships were also looked up using magic paths, which was a
cause of some fairly serious bugs.