Avoid re-saving records that haven't changed with batch imports.
This was already the case when using the "merge" rules, and used to
be in practice for all cases before we started allowing resaving of records
manually (the check was done closer to the database back then).

For a while now it has however been the case that when using --replaceBib
we would resave the record even without changes, which we should not.