Limit depth of frame/embed (#1125)
Limit depth of frame/embed.

For very connected datasets like Queerlit where many Concepts are related to many other Concepts in the same embellish graph there is a combinatorial explosion leading to huge framed documents (which are later thrown away).

We still want e.g. the queerlit concept scheme definition to be embedded in all places in the result so just add a limit on the number of levels where we start with a new "embedChain".