Search: Embellish with cards instead of chips in ES index
Fixes e.g. that instances can be found on contributor variant name
(LXL-3811).

This was previously changed from cards to chip because of problems with
the size of the ES index.

The switch on entity class (Item) should become vocab driven at some
point if we want to have different shapes like that.