Await threadpool shutdown in batchimport
Before starting work on the next import file.
Matches the behaviour of the old thread pool.

Imports of duplicates have been observerved in the wild when the import
job consists of a directory with many import files with a single record.
In this case it would of course be more efficient to not start a new
thread pool for every import file...