Use BlockingThreadPool instead of ThreadPool everywhere
One implementation of the blocking behaviour is enough.
Add a simpler way to use BlockingThreadPool with a single task queue.

Specifically in batchimport:
Same blocking behaviour as the old ThreadPool that was previously use here.
Don't queue up an unlimited amount of work. In case of extremely large import files.