CRUD API: Do all elasting indexing asynchronously
Hanging HTTP requests to elastic would cause POST or PUT to the CRUD API
to block for 60s (SO_TIMEOUT/READ_TIMEOUT in ElasticCLient). These would
timeout even if everything was finally successful leading clients to
retry and creating duplicates (in case of POST).

Do all elastic indexing in the background instead.

Reduce SO_TIMEOUT for indexing of single docs to 15s so that we retry
hanged requests sooner.