Change defaults for add-if-none priorities.
When merging impor records, the add-if-none rule allows the use of
a priority list, such that you're only allowed to add if your prio
is higher than the previous owners.

This changes your default priority (if you're not on the list), so
that you cannot add things. This avoids tedious need need for listing
every agent that is _not_ allowed to add this at this specific path.

If there is no prio list, the add-id-none rule will still allow anyone
to add at that path.