Search: Keep transformed + original strings in embellishments
When indexing, for things added by embellish we remove language
tagged string that are not in the whelk locale in order to keep
the index size down. With this change we also keep transformed strings
and their original counterpart (i.e. for now romanized + original script).
For example title of Instance linked from Item.