Search: Make filtering on multi-language properties work
Make filtering on for example prefLabel=foo also match
prefLabelByLang.sv, prefLabelByLang.en etc.

Quering prefLabelByLang.sv explicitly still works.

Add an extra field to the ES index that contains both the
language tagged and plain strings for a property.

For example:
```
{
  "foo": "FOO",
  "fooByLang": { "en": "EN", "sv": "SV" }
}
```
is indexed as
```
{
  "foo": "FOO",
  "fooByLang": { "en": "EN", "sv": "SV" },
  "__foo": ["FOO", "EN", "SV"]
}

```

Queries are adjusted accordingly to filter on this field.
Internal fields starting with '_' are already removed from the results.