Fix dependers handling in DatasetImporter (#1193)
Summary
-------
In DatasetImporter
* Make `--skip-dependers` flag only affect indexing
* Handle circular dependencies among deleted docs in dataset

Changes
-------
Remove the option to not update `lddb__dependencies` when loading a
dataset (`--skip-dependers`) since using it puts the database in an
inconsistent state. Updating `lddb_dependencies` is cheap because it is
only the _outgoing_ links for a modified doc that are updated.

`--skip-dependers` now only affects reindexing. Reindexing dependers
might be expensive when reloading core datasets. The flag can be
removed if we improve the logic for calculating which documents
need to be reindex for an update.

Handle circular dependencies among deleted docs in dataset. If no other
docs depend on the they can be deleted safely.