Add linking of serial editions to newspaper job
This adds to the exiting newspaper job currently running continuously to
amend newspaper data imported from Mimer.

The `editionStatement` of a newspaper issue is a code for a serial
edition described in the local Signe system (which has the intended
serial edition statement as its label).

The addition here adds an `isIssueOfEdition` link constructed from this
code, if such as edition exists in the system. It then removes the
editionStatement.

(Data from the Signe system is to be imported into Libris as a dataset.)