Attempt to run whelk-dependent integration tests
Some MarcFramePostProcStep:s are resource-dependent and need a backing
whelk to load those. This change adds a flag on those steps, which is
used to let MarcFrameConverterSpec conditionally run the specs for these
if it is able to instantiate a whelk upon startup (commonly when
integTest is run).

(This is conditional to make it possible to run the
non-resource-dependent steps locally without having a whelk up.)