Fix determination of content type from suffix
We inspect URI paths for suffixes, but do so too eagerly.

Formal request paths may contain dots (e.g. `etc` or `1.0`), so we need
to recognize the ones we use as representation paths.

We already use "data leafs" for this, i.e. the last path component is
special if it has a certain name (`data`, `find` or begins with an
underscore).

This fix makes sure only this representation data leaf path is used to
check for suffix-based content negotiation.

(And also refactors and simplifies some parts of the request inspection
where this is happening.)