Refactor history and fix shortcomings.
Before this change, the code had special treatment for the objects in the @graph
list, compared to all other lists.

This turned out to be a problem, because changes that (legitimately) needed to
claim ownership of these base objects didn't work right. This commit removes the
special treatment for these objects. This has consequences however, as the process
for claiming ownership had to be changed quite a bit to allow this.