Use Elasticsearch 8 (#1309)
*  Use Elasticsearch 8.8
* Add note about _es_id
* elasticCluster isn't actually ever used by the ES client, so remove it
* Update ES instructions in README
* ES index name: whelk_dev -> libris_local
* Increase number of shards and replicas

Changes necessary for XL to work with Elasticsearch 8. Main change is that ES 8 has security (and TLS) enabled by default, so we now have to user+pass over HTTPS when talking to the ES cluster (certificate generation and stuff like that is handled automatically by ES when starting for the first time).

The devops part has already been merged.

How to tune things for search speed is a separate issue that we can look at more once we have QA running in the new ES cluster. The increased number_of_shards / number_of_replicas is just a starting point based on our current usage in the main cluster and current recommendations (and anyway we should make number_of_shards and number_of_replicas configurable, but that's a devops thing).