Do linkFinder.normalizeIdentifiers in RemoteSearchAPI (#1344)
After conversion from MARC the documents might contain entities with
both `@id` and data

```
"language": { "@id": "https://id.kb.se/language/eng", "code": "eng" }
```

This trips up Embellisher which doesn't see it as a link (`@id` and size == 1).
So https://id.kb.se/language/eng is not added to embellish data.

The same thing then happens in the cataloging interface, so the
definition for English is not fetched there by `addMissingLinksToQuoted`
either. This leads to a chip with the URI instead of the label in the
form for the imported record.