Flatten CodeEquals-nodes in ASTs (optionally)
The point of this is to transform a query for example from:
somecode:(value1 or value2 and (value3 or value4))
into:
somecode:value1 or somecode:value2 and (somecode:value3 or somecode:value4))

The point of this is to reduce what can exist as a CodeEquals-operand to
strictly a String, which is then considerably easier to handle.

Effectively it turns the code:(group)-constuct into just syntactic sugar.